refactor(node-client): address review findings
- Reuse already-built credentials in connect() token branch
- Remove unnecessary Promise wrapper from close()
- Drop unused currentBackoff param from drainStream()
- Unexport and delete unused isAbortError function
- Extract shared sleep utilities to src/internal/sleep.ts
- Extract callbackToPromise helper to remove repeated promisification
