+++ clients/node/src/internal/sleep.ts
@@ -1,0 +1,27 @@
+export function sleep(ms: number, signal?: AbortSignal): Promise<void> {
+  return new Promise<void>((resolve, reject) => {
+    if (signal?.aborted) {
+      reject(new DOMException('Aborted', 'AbortError'))
+      return
+    }
+    const id = setTimeout(resolve, ms)
+    signal?.addEventListener('abort', () => {
+      clearTimeout(id)
+      reject(new DOMException('Aborted', 'AbortError'))
+    }, { once: true })
+  })
+}
+
+export function sleepUntilOrAbort(ms: number, signal: AbortSignal): Promise<boolean> {
+  return new Promise<boolean>((resolve) => {
+    if (signal.aborted) {
+      resolve(false)
+      return
+    }
+    const id = setTimeout(() => resolve(true), ms)
+    signal.addEventListener('abort', () => {
+      clearTimeout(id)
+      resolve(false)
+    }, { once: true })
+  })
+}
